style: :art: Move files to correct named folders
