+++ lib/screens/components/details_type.dart
