+++ lib/screens/components/poke_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:pokedex_project/helpers/constants.dart';
 import 'package:pokedex_project/helpers/text_functions.dart';
-import 'package:pokedex_project/pokemon/pokemon.dart';
+import 'package:pokedex_project/models/pokemon.dart';
 import 'package:pokedex_project/theme/card_color.dart';
 import 'package:pokedex_project/theme/type_color.dart';
