+++ lib/screens/home.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pokedex_project/helpers/wordings.dart';
-import 'package:pokedex_project/screens/home/components/pokemons_list.dart';
+import 'package:pokedex_project/screens/components/pokemons_list.dart';
 
 class Home extends StatelessWidget {
   const Home({super.key});
